Refactored Film Catalog and added new classes
Significant changes were made to the Film Catalog. The `Genre` property was removed from the `FilmCatalogResourceParameters.cs` class. The `DetailsFilmModel.cs` and `EditFilmViewModel.cs` classes were removed. The `FilmsPagedModel.cs` class was refactored to be a generic class with additional properties and methods. The `FilmsPropertyMapper.cs` class was refactored to include a new dictionary for property mapping and additional methods. The `FilmRepository.cs` class was refactored to include new methods for adding, deleting, getting, and updating films. The `IFilmRepository.cs` interface was updated to reflect these changes. The `DVDStore.Web.MVC.csproj` file was updated to remove the reference to the `Controllers` folder in the `FilmCatalog` area. A new suppression message was added to the `GlobalSuppressions.cs` file.

New files were added: `FilmsCatalogControllerTests.cs` for unit tests, `FilmsCatalogController.cs` for handling HTTP requests and responses, and `FilmViewModel.cs` to represent a Film in the MVC application.