Added role-based authorization to FilmsCatalogController
Implemented role-based authorization in the FilmsCatalogController using the Microsoft.AspNetCore.Authorization namespace. The [Authorize(Roles = "Manager")] attribute was added to the controller and all its action methods, restricting access to users in the "Manager" role. The Film Catalog navigation button in _Layout.cshtml was also modified to only be visible to "Manager" role users. Minor changes were made to the HTML and Bootstrap classes in _Layout.cshtml to modify the navbar behavior and appearance.