`Update Bootstrap stylesheets and admin dropdown color`
Updated the `_Layout.cshtml` file to include a variety of new Bootstrap stylesheets and removed a few old ones. The new stylesheets will provide a wider range of themes for the web application. Also added the `bootstrap-icons.min.css` stylesheet for additional icon styles. The background color for the "Administration" dropdown menu and the "User Administration" link has been changed to primary (`bg-primary`) to match the primary color defined in the Bootstrap stylesheet.